# Check and fix tensor and scalar type promotion