--- a/python/paddle/fluid/dygraph/math_op_patch.py
+++ b/python/paddle/fluid/dygraph/math_op_patch.py
@@ -162,6 +162,10 @@ def monkey_patch_math_varbase():
                 # in all cases(+, -, *, /, **, //, %), we need cast tensor.dtype to float
                 if self.dtype in _supported_int_dtype_:
                     self = astype(self, 'float32')
+                # here use `scale` replace `elementwise` to get better performance
+                # but only +, -, *, / can use this method
+                if scalar_method is not None:
+                    return scalar_method(self, other_var)
             elif isinstance(other_var, int):
                 # in all cases(+, -, *, /, **, //, %), we can cast it to float
                 # because the output tensor.dtype depend on the type of input tensor
@@ -169,17 +173,15 @@ def monkey_patch_math_varbase():
                 # division is a special case
                 if op_type == 'elementwise_div' and self.dtype in _supported_int_dtype_:
                     self = astype(self, 'float32')
+                # here use `scale` replace `elementwise` to get better performance
+                # but only +, -, *, / can use this method
+                if scalar_method is not None:
+                    return scalar_method(self, other_var)
             else:
                 # do nothing
                 pass
 
-            # 2. scalar method selected
-            # here use `scale` replace `elementwise` to get better performance
-            # but only +, -, *, / can use this method
-            if scalar_method is not None:
-                return scalar_method(self, other_var)
-
-            # 3. create varbase for scalar
+            # 2. create varbase for scalar
             lhs_dtype = self.dtype
             if not isinstance(other_var, core.VarBase):
                 if reverse:
@@ -189,7 +191,7 @@ def monkey_patch_math_varbase():
                     # add fill_op 
                     other_var = create_scalar(value=other_var, dtype=lhs_dtype)
 
-            # 4. unify right var type to left var
+            # 3. unify right var type to left var
             rhs_dtype = other_var.dtype
             if lhs_dtype != rhs_dtype:
                 other_var = astype(other_var, lhs_dtype)
@@ -198,7 +200,7 @@ def monkey_patch_math_varbase():
                 self = other_var
                 other_var = tmp
 
-            # 5. calculation
+            # 4. calculation
             axis = -1
             math_op = getattr(core.ops, op_type)
             return math_op(self, other_var, 'axis', axis)