… python/paddle/fluid/tests/unittests/test_tensor_scalar_type_promotion_static.py
… tools/__pycache__/static_mode_white_list.cpython-35.pyc